pyproject.toml
[tool.poetry]
name = "repospoofscanner"
version = "0.1.0"
description = "Scan GitHub repositories for malicious dependencies"
authors = ["Manikhilho"]
license = "MIT"
readme = "README.md"
packages = [{ include = "repospoofscanner" }]
[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.9.0"
rich = "^13.9.0"
httpx = "^0.27.0"
rapidfuzz = "^3.6.1"
PyGithub = "^1.59.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pre-commit = "^3.3.3"
mypy = "^1.5.1"
black = "^23.7.0"

[tool.poetry.scripts]
repospoof = "repospoofscanner.cli.main:app"